Struct piper::Mutex

pub struct Mutex<T> { /* fields omitted */ }

A mutex that implements async I/O traits.

This is a blocking mutex that adds the following impls:

impl<T> AsyncRead for Mutex<T> where &T: AsyncRead + Unpin {}

impl<T> AsyncRead for &Mutex<T> where &T: AsyncRead + Unpin {}

impl<T> AsyncWrite for Mutex<T> where &T: AsyncWrite + Unpin {}

impl<T> AsyncWrite for &Mutex<T> where &T: AsyncWrite + Unpin {}

This mutex is ensures fairness by handling lock operations in the first-in first-out order.

While primarily designed for wrapping async I/O objects, this mutex can also be used as a regular blocking mutex. It's not quite as efficient as parking_lot::Mutex, but it's still an improvement over std::sync::Mutex.

Examples

use futures::io;
use futures::prelude::*;
use piper::Mutex;

// Reads data from a stream and echoes it back.
async fn echo(stream: impl AsyncRead + AsyncWrite + Unpin) -> io::Result<u64> {
    let stream = Mutex::new(stream);
    io::copy(&stream, &mut &stream).await
}

Methods

impl<T> Mutex<T>

pub fn new(data: T) -> Mutex<T>

Creates a new mutex.

Examples

use piper::Mutex;

let mutex = Mutex::new(10);

pub fn lock(&self) -> MutexGuard<T>

Acquires the mutex, blocking the current thread until it is able to do so.

Returns a guard that releases the mutex when dropped.

Examples

use piper::Mutex;

let mutex = Mutex::new(10);
let guard = mutex.lock();
assert_eq!(*guard, 10);

pub fn try_lock(&self) -> Option<MutexGuard<T>>

Attempts to acquire the mutex.

If the mutex could not be acquired at this time, then None is returned. Otherwise, a guard is returned that releases the mutex when dropped.

Examples

use piper::Mutex;

let mutex = Mutex::new(10);
if let Ok(guard) = mutex.try_lock() {
    assert_eq!(*guard, 10);
}

pub fn into_inner(self) -> T

Consumes the mutex, returning the underlying data.

Examples

use piper::Mutex;

let mutex = Mutex::new(10);
assert_eq!(mutex.into_inner(), 10);

pub fn get_mut(&mut self) -> &mut T

Returns a mutable reference to the underlying data.

Since this call borrows the mutex mutably, no actual locking takes place -- the mutable borrow statically guarantees the mutex is not already acquired.

Examples

use piper::Mutex;

let mut mutex = Mutex::new(0);
*mutex.get_mut() = 10;
assert_eq!(*mutex.lock(), 10);